+++ lib/qr_code_dart_scan.dart
@@ -1,6 +1,9 @@
 library qr_code_dart_scan;
 
+export 'package:camera/camera.dart' show XFile;
 export 'package:zxing_lib/zxing.dart';
 
+export 'src/decoder/qr_code_dart_scan_decoder.dart';
+export 'src/qr_code_dart_scan_controller.dart';
 export 'src/qr_code_dart_scan_view.dart';
-export 'src/util.dart';
+export 'src/util/qr_code_dart_scan_resolution_preset.dart';